solved ValueError: Shapes (None, 1) and (None, 2) are incompatible
To solve I changed the attribute class_mode = "binary" this implies that the y_col "label" is a string and the class mode converts automatically to an integer. I made the mistake of converting labels to 1 and 0 only to get big errors. The second big issue in the ImageDataGenerator (.flow_from_generator) function. Also set up CNN with a sequential model that is actually working.